+++ gathering/services/configurations/base.py
