+++ gathering/services/configurations/exceptions.py
