+++ gathering/services/configurations/local_storage.py
