+++ gathering/services/ftp/collector.py
